changed the API to use routes instead of controllers under the advice of @Ducatel
